Address MR comments. Add tests.
