+++ docs/pages/quickstart/first_experiments/index.rst
@@ -9,4 +9,5 @@ commands to the robot every step), then evaluating a real pre-trained policy in 
    :maxdepth: 1
 
    exploring_variations
+   listing_variations
    running_a_real_policy/index
