+++ isaaclab_arena/evaluation/policy_runner.py
@@ -15,14 +15,11 @@ from isaaclab_arena.cli.isaaclab_arena_cli import get_isaaclab_arena_cli_parser
 from isaaclab_arena.evaluation.camera_video import CameraObsVideoRecorder
 from isaaclab_arena.evaluation.policy_runner_cli import add_policy_runner_arguments
 from isaaclab_arena.metrics.metrics_logger import metrics_to_plain_python_types
+from isaaclab_arena.utils.hydra_overrides import split_hydra_overrides
 from isaaclab_arena.utils.isaaclab_utils.simulation_app import SimulationAppContext
 from isaaclab_arena.utils.multiprocess import get_local_rank, get_world_size
 from isaaclab_arena.utils.random import set_seed
-from isaaclab_arena_environments.cli import (
-    get_arena_builder_from_cli,
-    get_isaaclab_arena_environments_cli_parser,
-    split_hydra_overrides,
-)
+from isaaclab_arena_environments.cli import get_arena_builder_from_cli, get_isaaclab_arena_environments_cli_parser
 
 if TYPE_CHECKING:
     from isaaclab_arena.policy.policy_base import PolicyBase
