+++ isaaclab_arena/tests/test_build_time_variations.py
@@ -102,45 +102,9 @@ def _test_enabled_build_time_variation_applied(simulation_app):
     return True
 
 
-def _test_hydra_override_applies_build_time_variation(simulation_app):
-    from isaaclab_arena.cli.isaaclab_arena_cli import get_isaaclab_arena_cli_parser
-    from isaaclab_arena.environments.arena_env_builder import ArenaEnvBuilder
-
-    # The variation starts disabled; the Hydra override must both enable it and set the
-    # sampler range, driving the radius to TEST_OVERRIDE_RADIUS (not the cfg default).
-    arena_env = get_test_environment(enabled=False)
-    sphere = arena_env.scene.assets[TEST_ASSET_NAME]
-    args_cli = get_isaaclab_arena_cli_parser().parse_args(["--num_envs", "1"])
-    builder = ArenaEnvBuilder(arena_env, args_cli)
-
-    builder.apply_hydra_variation_overrides([
-        f"{TEST_ASSET_NAME}.test_build_time.enabled=true",
-        f"{TEST_ASSET_NAME}.test_build_time.sampler_cfg.low=[{TEST_OVERRIDE_RADIUS}]",
-        f"{TEST_ASSET_NAME}.test_build_time.sampler_cfg.high=[{TEST_OVERRIDE_RADIUS}]",
-    ])
-
-    variation = sphere.get_variation("test_build_time")
-    assert variation.enabled, "Hydra override 'enabled=true' must enable the variation."
-
-    builder.compose_manager_cfg()
-
-    assert sphere.object_cfg.spawn.radius == pytest.approx(TEST_OVERRIDE_RADIUS, abs=1e-6), (
-        f"Hydra-overridden build-time variation must mutate '{TEST_ASSET_NAME}.object_cfg.spawn.radius' "
-        f"to {TEST_OVERRIDE_RADIUS}; got {sphere.object_cfg.spawn.radius}."
-    )
-    return True
-
-
 def test_disabled_build_time_variation_not_applied():
     assert run_simulation_app_function(
         _test_disabled_build_time_variation_not_applied,
-        headless=HEADLESS,
-    )
-
-
-def test_hydra_override_applies_build_time_variation():
-    assert run_simulation_app_function(
-        _test_hydra_override_applies_build_time_variation,
         headless=HEADLESS,
     )
